rename base auth strategy
