+++ src/authStrategies/BaseAuthStrategy.js
