+++ src/authStrategies/NoAuth.js
@@ -1,12 +1,12 @@
 'use strict';
 
-const AuthStrategy = require('./AuthStrategy');
+const BaseAuthStrategy = require('./BaseAuthStrategy');
 
 /**
  * No session restoring functionality
  * Will need to authenticate via QR code every time
 */
-class NoAuth extends AuthStrategy { }
+class NoAuth extends BaseAuthStrategy { }
 
 
 module.exports = NoAuth;
